+++ leggedsnake/__init__.py
@@ -25,7 +25,7 @@ from pylinkage import (
 from .utility import step, stride
 from .walker import Walker
 from .geneticoptimizer import (
-    evolutionary_optimization, genetic_optimization
+    evolutionary_optimization, GeneticOptimization
 )
 from .dynamiclinkage import (
     Nail, PinUp, DynamicPivot, Motor,
@@ -33,9 +33,13 @@ from .dynamiclinkage import (
     convert_to_dynamic_linkage,
 )
 from .physicsengine import (
-    video, video_debug,
+    video,
+    video_debug,
+    all_linkages_video,
     params,
-    World, VisualWorld
+    World,
+    VisualWorld
 )
+from .show_evolution import load_data, show_genetic_optimization
 
 __version__ = "0.3.1"