+++ leggedsnake/geneticoptimizer.py
@@ -34,7 +34,7 @@ def kwargs_switcher(arg_name, kwargs, default=None):
     """Simple function to return the good element from a kwargs dict."""
     out = default
     if arg_name in kwargs:
-        out = kwargs.pop(arg_name) or out
+        out = kwargs[arg_name] or out
     return out
 
 
@@ -254,11 +254,12 @@ def evolutionary_optimization_builtin(
 
     Parameters
     ----------
-    dna : list[float, tuple of float, tuple of tuple of float]
+    dna : list[float, tuple[float], tuple[tuple[float]]]
         DNA of individuals.
-    prob : list of float
-        List of probabilities of the good transmission of one
-        characteristics.
+    prob : list[float] | float
+        Standard deviation from a perfect transmission of a gene.
+        * float: standard deviation for all genes
+        * list[float]: standard deviation for each gene
     fitness : callable
         Evaluation function.
         Return float.
@@ -279,30 +280,29 @@ def evolutionary_optimization_builtin(
         startnstop : bool, default=False
             Ability to close program without loosing population.
             If True, we verify at initialization the existence of a data file.
-            Population is save every int(250 / max_pop) iterations.
+            Population is saved every int(250 / max_pop) iterations.
             The default is False.
         fitness_args : tuple
             Keyword arguments to send to the fitness function.
             The default is None (no argument sent).
         verbose : int
             Level of verbosity.
-            0 : no verbose, do not print anything.
-            1 : show a progress bar.
-            2 : complete report for each turn.
+            0: no verbose, do not print anything.
+            1: show a progress bar.
+            2: complete report for each turn.
             The default is 1.
         processes : int, default=1
             Number of processes that will evaluate the linkages.
 
     Returns
     -------
-    list[float, tuple of float, tuple of tuple of float]
+    list[float, tuple[float], tuple[tuple[float]]]
         List of 3-tuples: best score, best dimensions and initial positions.
         The list is sorted by score order.
     """
-    file_path = 'Population evolution.json'
     startnstop = kwargs_switcher('startnstop', kwargs, False)
-    if startnstop and os.path.exists(file_path):
-        pop = load_population(file_path)
+    if startnstop and os.path.exists(startnstop):
+        pop = load_population(startnstop)
     else:
         # At least two parents to begin with
         pop = [[dna[0], list(dna[1]), list(dna[2])] for _ in range(2)]
@@ -344,6 +344,8 @@ def evolutionary_optimization_builtin(
     for i in iterations:
         if verbose > 1:
             print(f"Turn: {i}, {len(pop)} individuals.")
+        if kwargs_switcher('gui', kwargs, False):
+            kwargs_switcher('gui', kwargs, False)(pop)
         # Population selection
         # Minimal score before death
         death_score = np.quantile([j[0] for j in pop], 1 - max_pop / len(pop))
@@ -352,13 +354,13 @@ def evolutionary_optimization_builtin(
         # We only keep max_pop individuals
         pop = list(filter(lambda x: x[0] >= death_score, pop))
         parents = select_parents(pop, verbose=verbose > 1)
-        # We select the best fit individual to show off, we know he is a parent
+        # We select the best fit individual to show off, we know it is a parent
         best_id = max(enumerate(parents), key=lambda x: x[1][0])[0]
         postfix[1] = parents[best_id][0]
         postfix[3] = parents[best_id][1]
         if startnstop:
             save_population(
-                file_path, pop, verbose > 1,
+                startnstop, pop, verbose > 1,
                 {
                     'best_score': parents[best_id][0],
                     'best_individual_id': best_id
@@ -377,7 +379,7 @@ def evolutionary_optimization_builtin(
 
     out = []
     for dna in pop:
-        # Return (fitness, dimensions, intial positions)
+        # Return (fitness, dimensions, initial positions)
         out.append(dna)
     out.sort(key=lambda x: x[0], reverse=True)
     return out
@@ -424,24 +426,24 @@ def evolutionary_optimization(
         startnstop : bool, optional
             Ability to close program without loosing population.
             If True, we verify at initialization the existence of a data file.
-            Population is save every int(250 / max_pop) iterations.
+            Population is saved every int(250 / max_pop) iterations.
             The default is False.
         fitness_args : tuple
             Positional arguments to send to the fitness function.
             The default is None (no argument sent).
         verbose : int, optional
             Level of verbosity.
-            0 : no verbose, do not print anything.
-            1 : show a progress bar.
-            2 : complete report for each turn.
+            0: no verbose, do not print anything.
+            1: show a progress bar.
+            2: complete report for each turn.
             The default is 1.
         processes : int, default=1
             Number of processes that will evaluate the linkages.
 
     Returns
     -------
-    list[dna]
-        An iterable of the best fit individuals, in format
+    list[float, tuple of float, tuple of tuple of float]
+        An iterable of the fittest individuals, in format
         (score, dimensions, initial coordinates).
 
     See Also
@@ -457,7 +459,7 @@ def evolutionary_optimization(
     )
 
 
-class genetic_optimization():
+class GeneticOptimization:
     def __init__(
             self, 
             dna,
@@ -465,7 +467,7 @@ class genetic_optimization():
             iters,
             prob=.07,
             **kwargs
-        ) -> None:
+    ) -> None:
         self.dna = dna
         self.fitness = fitness
         self.iters = iters
@@ -473,12 +475,8 @@ class genetic_optimization():
         self.kwargs = kwargs
 
     def run(self, iters, gui=False):
-        if gui:
-            raise NotImplementedError("Using gui=True not supported yet!")
-        else:
-            return evolutionary_optimization_builtin(
-                self.dna, self.prob, self.fitness,
-                iters=iters,
-                **self.kwargs
-            )
-
+        return evolutionary_optimization_builtin(
+            self.dna, self.prob, self.fitness,
+            iters=iters, gui=gui,
+            **self.kwargs
+        )